+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Fixed
 
+- Keep non-targeted members in the dependency-aware parallel template scheduler when another family member has release target branches.
+
 ### Security
 
 ## [1.2.79] - 2026-09-01